@@ -1,10 +1,10 @@
 name: relative_scale
 description: RelativeScale is a simple custom sizing system for flutter widgets to achieve the same physical sizes across different devices.
-version: 1.1.2
+version: 2.0.0
 homepage: https://github.com/xamantra/flutter_relative_scale
 
 environment:
-  sdk: ">=2.5.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
